make better

+++ almprm.hpp
@@ -1,5 +1,5 @@
 #include <vector>
 
-std::vector<int64_t> almprm1(const int, const int64_t);
-std::vector<int64_t> almprm2_1(const int, const int64_t);
-std::vector<int64_t> almprm2_2(const int, const int64_t);
+std::vector<int64_t> almprm1(const int8_t, const int64_t);
+std::vector<int64_t> almprm2_1(const int8_t, const int64_t);
+std::vector<int64_t> almprm2_2(const int8_t, const int64_t);
